Refactor: Added all activities inside activites folder
