+++ app/src/main/java/com/example/bookspresso/activities/AddBookActivity.java
@@ -1,4 +1,4 @@
-package com.example.bookspresso;
+package com.example.bookspresso.activities;
 
 import android.app.Activity;
 import android.content.Context;
@@ -8,7 +8,6 @@ import android.net.Uri;
 import android.os.Bundle;
 import android.provider.MediaStore;
 import android.view.MenuItem;
-import android.view.View;
 import android.widget.ArrayAdapter;
 import android.widget.Button;
 import android.widget.EditText;
@@ -20,6 +19,10 @@ import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.widget.Toolbar;
+
+import com.example.bookspresso.R;
+import com.example.bookspresso.database.DatabaseHelper;
+import com.example.bookspresso.models.Book;
 
 import java.text.SimpleDateFormat;
 import java.util.Date;
